INTERNAL: Prepare BTreeSMGet to clone operations.
