+++ src/main/java/net/spy/memcached/collection/BTreeSMGetWithByteTypeBkeyOld.java
@@ -18,167 +18,27 @@
 package net.spy.memcached.collection;
 
 import java.util.List;
-import java.util.Map;
 
-import net.spy.memcached.KeyUtil;
 import net.spy.memcached.MemcachedNode;
 import net.spy.memcached.util.BTreeUtil;
 
-public class BTreeSMGetWithByteTypeBkeyOld<T> implements BTreeSMGet<T> {
-
-  private static final String command = "bop smget";
-
-  protected MemcachedNode node;
-
-  protected String str;
-
-  protected List<String> keyList;
-  private String keySeparator;
-  private String spaceSeparatedKeys;
-
-  protected int lenKeys;
-
-  protected String range;
-  protected int offset = -1;
-  protected int count;
-  protected Map<Integer, T> map;
-
-  protected boolean reverse;
-
-  private String key;
-  private int flag;
-  private byte[] subkey;
-  private int dataLength;
-
-  private byte[] eflag = null;
-
-  private ElementFlagFilter eFlagFilter;
+public class BTreeSMGetWithByteTypeBkeyOld<T> extends BTreeSMGetImpl<T> {
 
   public BTreeSMGetWithByteTypeBkeyOld(MemcachedNode node, List<String> keyList,
                                        byte[] from, byte[] to,
                                        ElementFlagFilter eFlagFilter,
                                        int offset, int count) {
-    this.node = node;
-    this.keyList = keyList;
-    this.range = BTreeUtil.toHex(from) + ".." + BTreeUtil.toHex(to);
-    this.eFlagFilter = eFlagFilter;
-    this.offset = offset;
-    this.count = count;
-    this.reverse = BTreeUtil.compareByteArraysInLexOrder(from, to) > 0;
+    super(node, keyList, BTreeUtil.toHex(from) + ".." + BTreeUtil.toHex(to),
+        (BTreeUtil.compareByteArraysInLexOrder(from, to) > 0), eFlagFilter, offset, count);
   }
 
-  public void setKeySeparator(String keySeparator) {
-    this.keySeparator = keySeparator;
-  }
-
-  public String getSpaceSeparatedKeys() {
-    if (spaceSeparatedKeys != null) {
-      return spaceSeparatedKeys;
-    }
-
-    StringBuilder sb = new StringBuilder();
-    int numkeys = keyList.size();
-    for (int i = 0; i < numkeys; i++) {
-      sb.append(keyList.get(i));
-      if ((i + 1) < numkeys) {
-        sb.append(keySeparator);
-      }
-    }
-    spaceSeparatedKeys = sb.toString();
-    return spaceSeparatedKeys;
-  }
-
-  public MemcachedNode getMemcachedNode() {
-    return node;
-  }
-
-  public List<String> getKeyList() {
-    return keyList;
-  }
-
-  public String stringify() {
-    if (str != null) {
-      return str;
-    }
-
-    StringBuilder b = new StringBuilder();
-
-    b.append(KeyUtil.getKeyBytes(getSpaceSeparatedKeys()).length);
-    b.append(" ").append(keyList.size());
-    b.append(" ").append(range);
-
-    if (eFlagFilter != null) {
-      b.append(" ").append(eFlagFilter.toString());
-    }
-
-    if (offset > 0) {
-      b.append(" ").append(offset);
-    }
-
-    b.append(" ").append(count);
-
-    str = b.toString();
-    return str;
-  }
-
-  public String getCommand() {
-    return command;
-  }
-
-  public boolean headerReady(int spaceCount) {
-    return headerCount == spaceCount;
-  }
-
-  public String getKey() {
-    return key;
-  }
-
-  public int getFlag() {
-    return flag;
-  }
-
+  @Override
   public byte[] getSubkey() {
-    return subkey;
+    return (byte[]) subkey;
   }
 
-  public int getDataLength() {
-    return dataLength;
-  }
-
-  public boolean isReverse() {
-    return reverse;
-  }
-
-  public boolean hasEflag() {
-    return eflag != null;
-  }
-
-  public void decodeItemHeader(String itemHeader) {
-    String[] splited = itemHeader.split(" ");
-
-    /*
-    with flag
-      VALUE 1
-      SMGetTest31 0 0x01 0x45464C4147 6 VALUE1
-      MISSED_KEYS 0
-      END
-
-    without flag
-      VALUE 1
-      SMGetTest31 0 0x01 6 VALUE1
-      MISSED_KEYS 0
-      END
-     */
-    this.key = splited[0];
-    this.flag = Integer.parseInt(splited[1]);
-    this.subkey = BTreeUtil.hexStringToByteArrays(splited[2].substring(2));
-
-    if (splited[3].startsWith("0x")) {
-      this.eflag = BTreeUtil.hexStringToByteArrays(splited[3].substring(2));
-      this.dataLength = Integer.parseInt(splited[4]);
-    } else {
-      this.eflag = null;
-      this.dataLength = Integer.parseInt(splited[3]);
-    }
+  @Override
+  protected Object decodeSubkey(String subkey) {
+    return BTreeUtil.hexStringToByteArrays(subkey.substring(2));
   }
 }
