+++ src/main/java/net/spy/memcached/protocol/ascii/BTreeSortMergeGetOperationImpl.java
@@ -256,7 +256,7 @@ public class BTreeSortMergeGetOperationImpl extends OperationImpl implements
 
     if (lookingFor == '\0' && readOffset == data.length && count < lineCount) {
       BTreeSortMergeGetOperation.Callback cb = (BTreeSortMergeGetOperation.Callback) getCallback();
-      cb.gotData(smGet.getKey(), smGet.getSubkey(), smGet.getFlag(), data);
+      cb.gotData(smGet.getKey(), smGet.getSubkey(), smGet.getFlags(), data);
       lookingFor = '\r';
     }
