+++ src/main/java/net/spy/memcached/protocol/ascii/BTreeSortMergeGetOperationOldImpl.java
@@ -219,7 +219,7 @@ public class BTreeSortMergeGetOperationOldImpl extends OperationImpl implements
     if (lookingFor == '\0' && readOffset == data.length) {
       BTreeSortMergeGetOperationOld.Callback cb =
           (BTreeSortMergeGetOperationOld.Callback) getCallback();
-      cb.gotData(smGet.getKey(), smGet.getSubkey(), smGet.getFlag(), data);
+      cb.gotData(smGet.getKey(), smGet.getSubkey(), smGet.getFlags(), data);
       lookingFor = '\r';
     }
